requirements.txt
opencv-python
numpy
pyyaml
pynput
fastapi
uvicorn[standard]
websockets